server/requirements.txt
RPi.GPIO==0.7.1
opencv-python==4.8.1.78
numpy==1.26.2 